Added `six` 2/3 compatibility library
I used the [python-modernize]() script to do most of the work.  After cleaning up the `import`s (for the usual order—standard libs, third party, local), there was one error remaining in `runner.py`.

The error lies in the difference between how `filter()` works between different Python versions (and the fact that `six` imports a version of `filter` that works like Python 3).

Basically: in Python 2 `filter` usually returns a `list`.  In Python 3, it returns an iterator object.

So the fix was simple: wrap the statement in a list comprehension.  After that, the tests passed! :D

An (untested) alternative way to write the same statement would be:

```python

context_members = [i for i in inspect.getmembers(type(context_instance)) if not (i[0] in special_names or i[0].startswith('_'))]

```

…But I’m going to commit it with the contained `filter()`, because that is more similar to original code.