Refactored `VowsDefaultReporter` into different reporter classes…
…sort of.

Basically, this is intended to be “step 1” of more refactoring.  This doesn’t really change that much, because I just recombined the separate classes back into `VowsDefaultReporter` at the end of the file.

However, this does make it **much** easier to understand which pieces of code do what.

(**Also:** There were way too many printing methods. So, I changed several of them into string-processing methods, and then used regular `print` on the result.  This makes it easier to pass a string through many of these methods before printing it.)