Added file assertions (`assertions.types.file`)
I often find myself using files in my vows.  I thought I would cover this use case with a couple of new assertions. :)

But…there’s more than one way to check whether something is a “file”.  So how to write the assertion?

Well, one of PyVows’ strengths is the convenience of its built-in assertions.  The `to_be_like()` assertions are particularly fantastic.

In the spirit of such “fuzzy” assertions, I wrote the new file assertions with a touch of “fuzzy” logic as well.  If passed a string, the assertion checks whether the file exists.  If passed anything else, the assertion checks whether the topic is an instance of Python’s built-in file type.

I hope it’s useful!