Added `tests/assertions/file_vows.py`
Even with a testing syntax as awesome as PyVows, I’m not the greatest test writer. :/

But, I think this is a minimal start.  Room for future improvement includes:

- Add more test data
- In test context `WhenWeInstantiateThemAsFileObjects`, we need to make sure that open file objects are closed when that context is finished testing

On that second point, I looked into the `setup()` and `teardown()` methods.  I wanted to use `teardown()` to close the topic (an open file object)...but I couldn’t figure out how.

That’s because I have no way to pass the topic to `teardown()`.  If I want to close the topic when that context completes, `teardown()` seems like the obvious place to do it—but the topic needs to be in scope in order for that to happen.

Dammit.  ☹